Add details re: dependency version error when publishing packages
When publishing a package (with `sui client publish`) there
might be an error if the revision of a dependency used to
build the package does not match with the on-chain version.

Add details about the error and information about the
the revision (`rev`) field of a dependency that
appears in a package's `Move.toml`.